#2:  Add option for ignoring some dynamic parts during assertions
